Move shared filters to struct
